@@ -1,12 +1,10 @@
-# encoding: UTF-8
-
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
   s.name        = 'solidus_variant_picker'
-  s.version     = '2.2.0'
+  s.version     = '2.4.0'
   s.summary     = 'Enhanced variant listing on Spree product pages'
   s.description = 'Enhanced variant listing on Spree product pages'
-  s.required_ruby_version = '>= 2.2.2'
+  s.required_ruby_version = '>= 2.3'
 
   s.author = '@thogg4, @n00b2pr0, @ericsaupe'
   # s.email     = 'you@example.com'
@@ -17,19 +15,19 @@ Gem::Specification.new do |s|
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  s.add_dependency 'solidus', ['>= 1.0', '< 3']
   s.add_dependency 'deface'
+  s.add_dependency 'solidus', ['>= 1.0', '< 3']
 
+  s.add_development_dependency 'byebug'
   s.add_development_dependency 'capybara'
   s.add_development_dependency 'coffee-rails'
   s.add_development_dependency 'database_cleaner'
-  s.add_development_dependency 'factory_girl'
+  s.add_development_dependency 'factory_bot'
   s.add_development_dependency 'ffaker'
-  s.add_development_dependency 'rspec-rails'
-  s.add_development_dependency 'sass-rails'
   s.add_development_dependency 'poltergeist'
+  s.add_development_dependency 'rspec-rails'
+  s.add_development_dependency 'rubocop'
+  s.add_development_dependency 'sass-rails'
   s.add_development_dependency 'simplecov'
   s.add_development_dependency 'sqlite3'
-  s.add_development_dependency 'rubocop'
-  s.add_development_dependency 'pry'
 end
